move stuff from util to util
